refactor: move query params functions to utils
